+++ docs/about/release-notes.rst
@@ -35,12 +35,14 @@ Release Notes
    Jan-Lukas Wynen :sup:`a`\ ,
    and Sunyoung Yoo :sup:`a`
 
-
-vrelease
+vRELEASE_PLACEHOLDER
 --------
 
 Features
 ~~~~~~~~
+
+* :class:`scipp.Dataset` now supports ``assign_coords``, which updates or inserts the given coordinates to the :class:`scipp.Dataset` `#3110 <https://github.com/scipp/scipp/pull/3110/>`_.
+* :class:`scipp.DataArray` now supports ``assign_coords``/ ``assign_masks``/ ``assign_attrs``, which updates or inserts the given coordinates/masks/attributes to the :class:`scipp.DataArray` `#3110 <https://github.com/scipp/scipp/pull/3110/>`_.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~